feat (interceptor.js): translation of comments fix (UserService.cs): implementation of validation for the existence of the user's image.

+++ Todo.Application/Services/UserService.cs
@@ -47,7 +47,11 @@ public class UserService : IUserService
             await _todoItemRepository.DeleteByCreatedUserIdAsync(id);
         }
 
-        await _userPictureRepository.DeleteAsync(user.Picture);
+        if(user.Picture != null)
+        {
+            await _userPictureRepository.DeleteAsync(user.Picture);
+        }
+
         await _userRepository.DeleteAsync(user);
     }
